requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
passlib[bcrypt]
python-jose
langchain
chromadb
tiktoken
ollama
langchain-ollama
langchain-extensions